% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SDMengine.R
\name{make_progress_database}
\alias{make_progress_database}
\title{make_progress_database}
\usage{
make_progress_database(workshop_dir, configure)
}
\arguments{
\item{workshop_dir}{Directory of workshop}

\item{configure}{Configure object}
}
\description{
make_progress_database
}

