% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SDMengine.R
\name{query_progress_from_database}
\alias{query_progress_from_database}
\title{query_progress_from_database}
\usage{
query_progress_from_database(workshop_dir)
}
\arguments{
\item{workshop_dir}{Directory of workshop}
}
\description{
query_progress_from_database
}

