\name{sdmengine.dismo-package}
\alias{sdmengine.dismo-package}
\alias{sdmengine.dismo}
\docType{package}
\title{
dismo version implement of sdmegine
}
\description{
Using dismo(http://cran.r-project.org/web/packages/dismo/) as base engine to implement sdmengine
}
\details{
\tabular{ll}{
Package: \tab sdmengine.dismo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-17\cr
License: \tab AGPL\cr
}
}
\author{
Xiaoquan Kong,
Renyan Duan,
Minyi Huang

Maintainer: Xiaoquan Kong <u1mail2me@gmail.com>
}

\keyword{ species distribution modelling }
