% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sdmengine.dismo.implement.R
\name{sdmengine.dismo.implement}
\alias{sdmengine.dismo.implement}
\title{sdmengine.dismo.implement}
\usage{
sdmengine.dismo.implement(workshop_dir)
}
\arguments{
\item{workshop_dir}{directory of workshop}
}
\description{
entry interface of implement
}

