% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SDMengine.R
\name{SDMengine}
\alias{SDMengine}
\title{Main entry}
\usage{
SDMengine(workshop_dir, configure)
}
\arguments{
\item{workshop_dir}{Directory of workshop}

\item{configure}{Configure object}
}
\description{
Main entry
}

