% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/signal_worker.R
\name{signal_worker}
\alias{signal_worker}
\title{One CPU worker}
\usage{
signal_worker(parameter, workshop_dir, configure)
}
\arguments{
\item{parameter}{parameter of run}

\item{workshop_dir}{Directory of workshop}

\item{configure}{Configure object}
}
\description{
One CPU worker
}

